Color code todo-categories, create mat-icons for each todo dependent on status
Close #15 Categories are now color coded by category. Purple is the fallback color in the event that a un-predefined category is inserted.
Close #13 References to body content is in html, no need for any special CSS.
Close #7 mat-icons for true and false exist
Close #1 We are satisfied with how todos are added.
Close #4 Add todo component has valid html